Replaced my complicated "evolved" linked list implementation with a slow but very simple implementation. likely far more reliable.
During this I also fixed the event problems which were caused when some events occured at the same time and got ordered last first. Now they are first first. Also the previous event is applied to the step immediatly after it. which is really weird, but works. More looking is needed.